backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
chromadb
pydantic
firebase-admin
python-dotenv
phidata
duckduckgo-search
langchain-community
sentence-transformers
pypdf
openai
groq
langchain-chroma
python-multipart

